requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
crypto==1.4.1
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-allauth==0.44.0
django-chartjs==2.2.1
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
mysqlclient==2.0.3
Naked==0.1.31
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
shellescape==3.8.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
